generator/requirements.txt
openai
pillow
python-dotenv
gradio